+++ examples/buffer_deletion.rs
@@ -4,7 +4,7 @@ fn main() {
     let app = app::App::default();
     let mut win = window::Window::new(100, 100, 800, 600, "Test");
     let mut buf = text::TextBuffer::default();
-    let _editor = text::TextEditor::new(0, 0, 800, 600, buf);
+    let mut _editor = text::TextEditor::new(0, 0, 800, 600, buf);
     win.show();
 
     win.set_callback(Box::new(move || {