+++ src/lib.rs
@@ -183,6 +183,7 @@
 //! - system-zlib: Uses the system zlib
 //! - legacy-opengl: Support of Lagacy OpenGL, the crate uses GLVND by default
 //! - fltk-bundled: Support for bundled versions of cfltk and fltk on selected platforms
+//! - no-opengl: Support for systems without OpenGL.
 //!
 //! ## FAQ
 //!