+++ src/prelude.rs
@@ -387,6 +387,25 @@ pub unsafe trait MenuExt: WidgetExt {
         flag: crate::menu::MenuFlag,
         cb: Box<dyn FnMut()>,
     );
+    /// Add a menu item along with an emit (sender and message)
+    fn add_emit<T: 'static + Copy + Send + Sync>(
+        &mut self,
+        name: &str,
+        shortcut: Shortcut,
+        flag: crate::menu::MenuFlag,
+        sender: crate::app::Sender<T>,
+        msg: T,
+    );
+    /// Inserts a menu item along with an emit (sender and message)
+    fn insert_emit<T: 'static + Copy + Send + Sync>(
+        &mut self,
+        idx: u32,
+        name: &str,
+        shortcut: Shortcut,
+        flag: crate::menu::MenuFlag,
+        sender: crate::app::Sender<T>,
+        msg: T,
+    );
     /// Remove a menu item by index
     fn remove(&mut self, idx: u32);
     /// Adds a simple text option to the Choice and MenuButton widgets