+++ src/widget.rs
@@ -38,8 +38,8 @@ impl Widget {
     }
 
     /// Transform Widget base to another Widget
-    pub fn into<W: WidgetExt>(self) -> W {
+    pub unsafe fn into<W: WidgetExt>(self) -> W {
         assert!(!self.was_deleted());
-        unsafe { W::from_widget_ptr(self._inner) }
+        W::from_widget_ptr(self._inner)
     }
 }